Added a name field in the Raspberries table.
This will allow users to name their raspberry pi's for easier
reference in the future (rather than having to type in the full
UUID). Also, changed the primary in the table to (raspberryID, userID)
in order to allow multiple users per raspberry. To deal with this
change (as raspberryID no longer uniquely identifies every entry
in the table), Events' foreign key for raspberryid was removed.