Added RaspberryNames table, removing names from Raspberries table.
After further discussion, the decision to add names to the Raspberries table
was reversed. Instead, we will be opting to create a new table called
RaspberryNames which will store the name linked to each raspberry. This
means that a raspberry will only ever have one 'name' linked to it at
any time. This decision came because Iman and I thought the raspberry
should be treated as a router; and also because of storage issues. The
RaspberryNames table will only ever be as large as there are raspberries,
while the Raspberries table has as many entries as there are users. If we
had 1000 users connected to 250 raspberries, we'd be storing 500 pieces of new
information in the new table, while the name column in Raspberries would require
1000 entries on its own.